==> 1.DFW_daily_flights_intro.ipynb <==
# INTRODUCTION
This notebook combines and explores daily flight data from Dallas Fort Worth (DFW) airport and daily weather data for DFW from Jan 1 2018 to Nov 17, 2023. The purpose is to produce models that can predict daily flight traffic, such as the number of total, ontime, cancelled, and delayed flights at Dallas Fort Worth Airport (DFW) from historical flight and weather data. This analysis is exploratory, but could be valuable to DFW airport management.

The daily weather data were obtained from a friend, Monte Lunacek, at the National Renewable Energy Laboratory (NREL) who obtained the data from the National Oceanic and Atmospheric Administration (NOAA), High-Resolution Rapid Refresh (HRRR) Data Archive: AWS OPen Data Program. The weather flight data was scaped from the Bureau of Transporation Statistics using Selenium.. 

In addition to merging the daily flight and weather data, this notebook engineers many features potentially relevant to flight traffic, such as time, date, and holiday features, and lag columns for total flights and cancelled flights. Column lists are created for convenient handling of features and targets. The final dataset is saved as "datav3/daily/daily_flights_and_weather_merged.parquet" for use in machine learning peformed in other notebooks.


## Libraries

import os
import json
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import matplotlib.dates as mdates
import holidays
import datetime as dt

import seaborn as sns
import sklearn
from sklearn.model_selection import train_test_split
from sklearn.linear_model import LinearRegression as LR
from sklearn.metrics import mean_squared_error, r2_score
from sklearn.preprocessing import PolynomialFeatures
from sklearn.pipeline import make_pipeline
from sklearn.preprocessing import StandardScaler
from sklearn.svm import SVR
from sklearn.ensemble import RandomForestRegressor
from sklearn.model_selection import GridSearchCV
from sklearn.compose import ColumnTransformer
from sklearn.preprocessing import OneHotEncoder, LabelEncoder, FunctionTransformer
from sklearn.decomposition import PCA

DAILY_DATA_PATH = "data.v3/daily" 

pd.set_option("display.width", 600)

# DAILY FLIGHT DATA

flights_d = pd.read_parquet(os.path.join(DAILY_DATA_PATH, "daily_flights.parquet"))
print(f"Daily flights columns: {flights_d.columns.tolist()}")
print('Daily flights import shape:', flights_d.shape)

## Daily flight data - initial look

flights_d.describe().round().T

flights_d.info()

flights_d.head().T

flights_d.tail().T

## Daily flight data - missing values

flights_d['date'] = flights_d.index
flights_d['year'] = flights_d['date'].dt.year
missing = flights_d.groupby('year').apply(lambda x: x.isnull().sum())
print("Missing values per year:", missing)

# DAILY WEATHER

## Daily weather data - source

Source: NOAA Global Systems Laboratory, High-Resolution Rapid Refresh (HRRR) Data Archive: AWS OPen Data Program\
Data dictionary: [HRRR Zarr Variable List](https://mesowest.utah.edu/html/hrrr/zarr_documentation/html/zarr_variables.html)

## Daily weather data - import 

weather_d = pd.read_parquet(os.path.join(DAILY_DATA_PATH, "daily_weather.parquet"))
print(f"Weather columns: {weather_d.columns.tolist()}")
print("Daily weather import shape:", weather_d.shape)

## Daily weather data - initial look

print("Weather null count and data types:")
print(weather_d.info())

print("\n\nWeather descriptives:")
print(weather_d.describe().round().T)

weather_d.head().T

weather_d.tail().T

## Daily weather - missing values

# count the missing values in each column aggregated by year
weather_d['date'] = weather_d.index
weather_d['year'] = weather_d['date'].dt.year
missing = weather_d.groupby('year').apply(lambda x: x.isnull().sum())
print("Missing values by year and variable\n", missing.T)

# Show rows where the year is 2022 and there are missing values in the weather data
print("\nAnomalous missing values on October 1st, 2022\n")
print(weather_d[(weather_d['year'] == 2022) & (weather_d.isnull().any(axis=1))])

## Daily weather - data cleaning

### Drop years and rows with no weather data

weather_d_incomplete = weather_d[weather_d.isnull().any(axis=1)].copy()
weather_d_complete = weather_d[~weather_d.isnull().any(axis=1)].copy()

print("Number of weather rows with missing values:", weather_d_incomplete.shape[0])
print("\nDaily weather final shape:", weather_d_complete.shape)

### Updated weather data head and tail

print("Head\n", weather_d_complete.head().T)
print("\n\nTail\n", weather_d_complete.tail().T)

# MERGE FLIGHT AND WEATHER DATA

Note: The weather data covers a shorter time period than the flights data. Can we get more weather data?

# Merge flights and weather data
df = pd.merge(flights_d.drop(['year', 'date'], axis=1), 
              weather_d_complete.drop(['year', 'date'], axis=1), 
              left_index=True, right_index=True, how='inner')

print("flights_d shape:", flights_d.shape)
print("weather_d_complete shape:", weather_d_complete.shape)
print("merged shape:", df.shape)
print('\nHEAD\n', df.head(2).T)
print('\nTAIL\n', df.tail(2).T)

# FEATURE ENGINEERING

## Add standardized and squared weather columns

weather_cols = ['wx_temperature_max', 'wx_temperature_min', 'wx_apcp', 'wx_prate', 'wx_asnow', 'wx_frozr', 'wx_vis', 'wx_gust', 'wx_maxref', 'wx_cape', 'wx_lftx', 'wx_wind_speed', 'wx_wind_direction']

for col in weather_cols:
    mean = df[col].mean()
    stdev = df[col].std()
    df.loc[:, f'{col}_s2'] = ((df[col] - mean)/stdev) ** 2

weather_cols_s2 = [ f"{col}_s2" for col in weather_cols]
print("Standardized squared weather columns added to the dataframe\n", df[weather_cols_s2].head().T)

## Add time, date, and holiday features

import datetime as dt
from dateutil.easter import easter

# create a date column from index
df['date'] = df.index
df['year'] = df['date'].dt.year
# df['year_c'] = df['date'].dt.year.astype('category')
df['month'] = df['date'].dt.strftime('%B')
df['day_of_week'] = df['date'].dt.strftime('%A')
df['day_of_month'] = df['date'].dt.day

# Create an ordinal date column (days since 1/1/1)
df['ordinal_date'] = df['date'].map(dt.datetime.toordinal)

# add season column
def get_season(date):
    if date.month in [12, 1, 2]:
        return 'winter'
    elif date.month in [3, 4, 5]:
        return 'spring'
    elif date.month in [6, 7, 8]:
        return 'summer'
    else:
        return 'fall'

df['season'] = df['date'].apply(get_season)

# Holidays
years = range(df['year'].min(), df['year'].max()+1)
us_holidays = holidays.US(years = years)

# Reverse dictionaries of US Holidays spanning the years in the dataset
# for year in years:
#     globals()[f"holidays_{year}"] = {v: k for k, v in holidays.US(years=year).items()}

# add holiday column
def check_holiday(date):
    # us_holidays = holidays.US(years = date.year)
    return us_holidays.get(date, "Not a Holiday")
df['holiday'] = df['date'].apply(check_holiday)

# add "halloween" column
def halloween(date):
    if date.month == 10 and date.day == 31:
        return 'yes'
    else:
        return 'no'
df['halloween'] = df['date'].apply(halloween)

# add "jan_2" column
def jan_2(date):
    if date.month == 1 and date.day == 2:
        return 'yes'
    else:
        return 'no'
df['jan_2'] = df['date'].apply(jan_2)
    
# add "jan_3" column
def jan_3(date):
    if date.month == 1 and date.day == 3:
        return 'yes'
    else:
        return 'no'
df['jan_3'] = df['date'].apply(jan_3)
    

# add column for the day before easter
def day_before_easter(date):
    normalized_input_date = date.date()
    day_before_easter_date = easter(date.year) - dt.timedelta(days=1)
    if normalized_input_date == day_before_easter_date:
        return 'yes'
    else:
        return 'no'
df['day_before_easter'] = df['date'].apply(day_before_easter)

# add column for the Xmas Eve
def xmas_eve(date):
    if date.month == 12 and date.day == 24:
        return 'yes'
    else:
        return 'no'
df['xmas_eve'] = df['date'].apply(xmas_eve)

# add column for New Year's eve
def new_years_eve(date):
    if date.month == 12 and date.day == 31:
        return 'yes'
    else:
        return 'no'
df['new_years_eve'] = df['date'].apply(new_years_eve)

# add "days_until_Xmas" column
def days_until_xmas(date):
    xmas = pd.to_datetime(f"{date.year}-12-25")
    if date > xmas:
        xmas = pd.to_datetime(f"{date.year+1}-12-25")
    return (xmas - date).days
df['days_until_xmas'] = df['date'].apply(days_until_xmas)


# add "days_until_thanksgiving" column
def days_until_thanksgiving(date):
    # Thanksgiving is the 4th Thursday of November
    # https://www.timeanddate.com/holidays/us/thanksgiving-day
    thanksgiving = pd.to_datetime(f"{date.year}-11-01")
    while thanksgiving.weekday() != 3:
        thanksgiving += pd.Timedelta(days=1)
    thanksgiving += pd.Timedelta(days=21)
    return (thanksgiving - date).days
df['days_until_thanksgiving'] = df['date'].apply(days_until_thanksgiving)

# add "days_until_July_4th" column
def days_until_july_4th(date):
    july_4th = pd.to_datetime(f"{date.year}-07-04")
    if date > july_4th:
        july_4th = pd.to_datetime(f"{date.year+1}-07-04")
    return (july_4th - date).days
df['days_until_july_4th'] = df['date'].apply(days_until_july_4th)

# add "days_until_labor_day" column
def days_until_labor_day(date):
    # Labor Day is the first Monday of September
    labor_day = pd.to_datetime(f"{date.year}-09-01")
    while labor_day.weekday() != 0:
        labor_day += pd.Timedelta(days=1)
    return (labor_day - date).days
df['days_until_labor_day'] = df['date'].apply(days_until_labor_day)

# add "days_until_memorial_day" column
def days_until_memorial_day(date):
    # Memorial Day is the last Monday of May
    memorial_day = pd.to_datetime(f"{date.year}-05-01")
    if date > memorial_day:
        memorial_day = pd.to_datetime(f"{date.year+1}-05-01")
    return (memorial_day - date).days
df['days_until_memorial_day'] = df['date'].apply(days_until_memorial_day)

# add covid column. The WHO declared COVID-19 a pandemic on March 11, 2020. Record-breaking Thanksgiving travel in 2023 suggests a full recovery by 11/23/23.
mask = (df['date'] >= '2020-03-11') & (df['date'] <= '2023-11-23')
df['covid'] = np.where(mask, 'yes', 'no')

# Move date-related variables to front of dataframe
cols = df.columns.tolist()
date_cols= ['date', 'covid', 'ordinal_date', 'year', 'month', 'day_of_month', 'day_of_week', 'season', 'holiday', 'halloween', 'xmas_eve', 'new_years_eve', 'jan_2', 'jan_3', 'day_before_easter', 'days_until_xmas', 'days_until_thanksgiving', 'days_until_july_4th', 'days_until_labor_day', 'days_until_memorial_day']
col_order = date_cols + [col for col in cols if col not in date_cols]
df = df[col_order]


# Verify the day before easter dates make sense
print("Day before easter dates:", df[df['day_before_easter'] == 'yes']['date'].tolist())

# Preview the dataframe
print("\nDate-related columns:\n")
df[date_cols].head().T

## Add flights total and percent delayed, ontime, cancelled columns

# Calculate percentages for total flights
df['flights_total'] = df['flights_cancel'] + df['flights_delay'] + df['flights_ontime']
df['flights_cancel_pct'] = 100*df['flights_cancel'] / df['flights_total']
df['flights_delay_pct'] = 100*df['flights_delay'] / df['flights_total']
df['flights_ontime_pct'] = 100*df['flights_ontime'] / df['flights_total']

# Calculate percentages for arrivals
df['flights_arr_delay_pct'] = 100*df['flights_arr_delay'] / df['flights_arr']
df['flights_arr_ontime_pct'] = 100*df['flights_arr_ontime'] / df['flights_arr']
df['flights_arr_cancel_pct'] = 100*df['flights_arr_cancel'] / df['flights_arr']

# Calculate percentages for departures
df['flights_dep_delay_pct'] = 100*df['flights_dep_delay'] / df['flights_dep']
df['flights_dep_ontime_pct'] = 100*df['flights_dep_ontime'] / df['flights_dep']
df['flights_dep_cancel_pct'] = 100*df['flights_dep_cancel'] / df['flights_dep']

print(df.head().T)

print('shape:', df.shape)

## Add 7 lag features each for cancelled, delayed, and ontime flights using 1 day steps
The lag columns will be used exclusively as features. Because total flights can be derived from ontime, delayed, and cancelled flights, we will not create a lag column for total flights.

lag_cols = []

cols_to_lag = ['flights_cancel', 'flights_delay', 'flights_ontime']
for col in cols_to_lag:
    for lag in [1, 2, 3, 4, 5, 6, 7]:
        df[f"{col}_lag_{lag}"] = df[col].shift(lag)
        lag_cols.append(f"{col}_lag_{lag}")

df.dropna(inplace=True)

print(df.shape)
print(f"Lag cols: {lag_cols}")
print("\n", df[['flights_cancel', 'flights_cancel_lag_1', 'flights_cancel_lag_2', 'flights_cancel_lag_3']].head().T)
print("\n", df[['flights_delay', 'flights_delay_lag_1', 'flights_delay_lag_2', 'flights_delay_lag_3']].head().T)
print("\n", df[['flights_ontime', 'flights_ontime_lag_1', 'flights_ontime_lag_2', 'flights_ontime_lag_3']].head().T)

## Add a random feature to test RFE

# Add a random variable to df_lag with repeatable results
np.random.seed(42)
df['random'] = np.random.randint(0, 100, df.shape[0])

# GROUPED COLUMN LISTS (flights, weather, date)

# Flights column groups
flights_terminal_cols = ['flights_arr_A', 'flights_arr_B', 'flights_arr_C', 'flights_arr_D', 'flights_arr_E',
                         'flights_dep_A', 'flights_dep_B', 'flights_dep_C', 'flights_dep_D', 'flights_dep_E']

flights_non_terminal_cols = ['flights_total', 'flights_cancel', 'flights_delay', 'flights_ontime',
                             'flights_arr_ontime', 'flights_arr_delay', 'flights_arr_cancel',
                             'flights_dep_ontime', 'flights_dep_delay', 'flights_dep_cancel']

flights_percentage_cols = ['flights_cancel_pct', 'flights_delay_pct', 'flights_ontime_pct',
                            'flights_arr_delay_pct', 'flights_arr_ontime_pct', 'flights_arr_cancel_pct',
                            'flights_dep_delay_pct', 'flights_dep_ontime_pct', 'flights_dep_cancel_pct']

flights_prediction_cols = flights_non_terminal_cols + flights_percentage_cols
flights_forecast_cols = [f"{col}_next_day" for col in flights_prediction_cols]

# Date column groups
date_cols = ['date', 'covid', 'ordinal_date', 'year', 'month', 'day_of_month', 'day_of_week', 'season', 'holiday', 'halloween', 'xmas_eve', 'new_years_eve', 'jan_2', 'jan_3', 'day_before_easter', 'days_until_xmas', 'days_until_thanksgiving', 'days_until_july_4th', 'days_until_labor_day', 'days_until_memorial_day']

# Weather column groups
weather_cols = ['wx_temperature_max', 'wx_temperature_min', 'wx_apcp', 'wx_prate', 'wx_asnow', 'wx_frozr', 'wx_vis', 'wx_gust', 'wx_maxref', 'wx_cape', 'wx_lftx', 'wx_wind_speed', 'wx_wind_direction']
weather_cols_s2 = ['wx_temperature_max_s2', 'wx_temperature_min_s2', 'wx_apcp_s2', 'wx_prate_s2', 'wx_asnow_s2', 'wx_frozr_s2', 'wx_vis_s2', 'wx_gust_s2', 'wx_maxref_s2', 'wx_cape_s2', 'wx_lftx_s2', 'wx_wind_speed_s2', 'wx_wind_direction_s2']

# Lag column groups
lag_cols =  ['flights_cancel_lag_1', 'flights_cancel_lag_2', 'flights_cancel_lag_3', 'flights_cancel_lag_4', 'flights_cancel_lag_5', 'flights_cancel_lag_6', 'flights_cancel_lag_7',
             'flights_delay_lag_1', 'flights_delay_lag_2', 'flights_delay_lag_3', 'flights_delay_lag_4', 'flights_delay_lag_5', 'flights_delay_lag_6', 'flights_delay_lag_7',
             'flights_ontime_lag_1', 'flights_ontime_lag_2', 'flights_ontime_lag_3', 'flights_ontime_lag_4', 'flights_ontime_lag_5', 'flights_ontime_lag_6', 'flights_ontime_lag_7']

# ADD FORECAST COLUMNS (targets shifted one day)

for target in flights_prediction_cols:
    df[f"{target}_next_day"] = df[target].shift(-1) # next day's value

# EXPORT DATA TO PARQUET

# Export the final dataframe
df.to_parquet(os.path.join(DAILY_DATA_PATH, "daily_flights_and_weather_merged.parquet"))
